+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 - Add `class` attribute to `:item` slot of `Doggo.navbar_items/1`.
 - Add `required_title` attribute to `Doggo.label/1`.
 - Change `description` attribute of `Doggo.field_description/1` to inner block.
+- Require `id` attribute for `Doggo.drawer_section/1`.
+- Add `aria-labelledby` attribute to `Doggo.drawer_section/1`.
 
 ### Fixed